Switch from openstacksdk library to openstackclients snap
The `openstacksdk` library depends on some crypto libraries which end up
requiring compilation during `charm build` and this causes errors when
using the `charm` snap. This switches to the CLI `openstackclients` snap
instead to avoid that issue.

This also adds checking for `auth-type` which, while not addressing the
underlying problem, should at least provide a better user experience for
https://bugs.launchpad.net/charm-openstack-integrator/+bug/1835768