Handle cached creds not containing version
We previously ignored the version from the creds (if available), but #18
added handling for that.  The cached creds from before the upgrade did
not contain it and caused a KeyError.

Also updated the description of `bs-version` to call out that it is
specifically a block storage setting, rather than an override for the
API version generally.

Fixes https://bugs.launchpad.net/charm-openstack-integrator/+bug/1841271