Fix API version not being detected correctly in some cases
It seems that the openstackclients CLI doesn't always properly detect
the version when it is not explicitly specified, so we need to do it
ourselves. This also makes errors while trying to detect Octavia
non-fatal and prevents useful error messages from being hidden from the
logs.

Fixes https://bugs.launchpad.net/charm-openstack-integrator/+bug/1841607