Fix octavia detection logic when octavia exists on multiple regions (#53)
PR 49 [1] updated the Octavia detection logic to consider region
name. However the change checks if region is same only for first
endpoint. In case of multiple regions deploying Octavia, this will
fail.

The fix updates the Octavia detection logic to verify if any one of
the endpoint matches the region name.

Fixes lp:1947085

[1] https://github.com/juju-solutions/charm-openstack-integrator/pull/49

Co-authored-by: Hemanth Nakkina <hemanth.nakkina@canonical.com>