Add opinionated health monitor for loadbalancers
This loadbalancer integration is used by kubernetes master,
but previous did not include any health checks,
which caused issues with managing the member list.
So, add a health monitor to fix the issue.

It must be a TLS-HELLO monitor; we can't use the /livez k8s endpoint,
because it requires authentication.
This should be good enough to monitor port 6443
of the kubernetes master service.

Note that we can be opinionated here,
because this load balancer integration is only
used with charmed kubernetes.